Start instance ok in backend

+++ backend/models/docker_config.go
@@ -1,11 +1,11 @@
 package models
 
 type DockerConfig struct {
-	ID               uint   `gorm:"primaryKey" json:"id"`
-	Host             string `json:"host"`
-	ImagePrefix      string `json:"imagePrefix"`
-	InstancesByTeam  int    `json:"instancesByTeam"`
-	InstancesByUser  int    `json:"instancesByUser"`
-	MaxMemByInstance int    `json:"maxMemByInstance"`
-	MaxCpuByInstance int    `json:"maxCpuByInstance"`
+	ID               uint    `gorm:"primaryKey" json:"id"`
+	Host             string  `json:"host"`
+	ImagePrefix      string  `json:"imagePrefix"`
+	InstancesByTeam  int     `json:"instancesByTeam"`
+	InstancesByUser  int     `json:"instancesByUser"`
+	MaxMemByInstance int     `json:"maxMemByInstance"`
+	MaxCpuByInstance float64 `json:"maxCpuByInstance"`
 }
